requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.4.247
cryptography==44.0.0
distro==1.9.0
fake-http-header==0.3.5
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.59.10
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
nltk==3.9.1
numpy==2.2.2
openai==1.60.2
packaging==24.2
pillow==10.4.0
playwright==1.49.1
propcache==0.2.1
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.0.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
setuptools==75.8.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
tf-playwright-stealth==1.1.1
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
